더보기 페이지 구현(로그아웃)
